refactor: reorganize project structure and update configuration files. Zero-3rdparty package tests ok and build ok
