+++ zero-3rdparty/pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "zero-3rdparty"
+version = "1.0.0+rc1"
+requires-python = ">=3.10"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.sdist]
+include = [
+  "zero_3rdparty/*.py",
+]
+# exclude = [
+#   "*.json",
+#   "pkg/_compat.py",
+# ]
+
+#  [tool.hatch.build.targets.wheel]
+#  packages = ["src/zero_3rdparty"]
