+++ zero-3rdparty/zero_3rdparty/decorator_globals.py
