+++ zero-3rdparty/zero_3rdparty/future.py
@@ -11,9 +11,7 @@ from concurrent.futures import Future as _ConcFuture
 from concurrent.futures import TimeoutError as _ConcTimeoutError
 from contextlib import suppress
 from functools import wraps
-from typing import Any, Callable, TypeVar, Union
-
-from typing_extensions import TypeAlias
+from typing import Any, Callable, TypeVar, TypeAlias, Union
 
 logger = logging.getLogger(__name__)
 ResultT = TypeVar("ResultT")
