+++ zero-3rdparty/zero_3rdparty/iter_utils.py
