--- zero_3rdparty/tests/test_zero_3rdparty/BUILD
@@ -1,1 +1,0 @@
-python_tests(**test_params())
